+++ CHANGELOG.md
@@ -4,7 +4,6 @@
  - Torch inputs to `cwt` and `ssq_cwt` now supported
 
 #### FIXES
-
  - `_gmw.mom2cum`: fix formula mistake, `range(1, n - 1)` -> `range(1, n)`, by @bartvm
  - fix flipping `ssq_freqs` in torch
  - `istft` edge case with `dtype='float32'`, time-localized `window`, and large `hop_len`
@@ -14,9 +13,8 @@
  - fixed important typo in `ridge_extraction.py`
  - fix: librosa defaults to zero padding, which broke `reconstruction_test`
  - fixed `nv` criterion for ignoring sanity check
-
+ 
 #### MISC 
-
  - slight performance boost in `_process_ssq_params`
  - extended `visuals.hist`
  - removed `xp -= xp.mean()`, a no-op with CWT, badly copied from MATLAB (original ssq code)
@@ -24,6 +22,7 @@
  - simplify `utils.backend.astype`
  - simplify `p2up`
  - `padsignal` supports torch inputs
+ - `nan_checks` now defaults to `True` only for NumPy inputs
 
 
 ### 0.6.3 (1-23-2022): QoL, cleanups, fixes